backend/requirements.txt
flask
flask-cors
pillow
google-cloud-vision
googletrans==4.0.0-rc1
